fix: widen the order number and stop a cart quantity wrapping negative
Two defects found reading the checkout path after the hardening pass.

Order numbers collided at commerce scale. The number is WW-{date}-{suffix} and
the suffix was six hex characters of the order's Guid - 24 bits, scoped to a
single day. order_number carries a unique index, so a collision was never a
data leak, but it was an INSERT that violated the constraint and rolled the
whole placement back: a customer meeting a hard failure at checkout.

Collisions arrive by the birthday bound, not when the space runs out, so this
bites far earlier than 16.7 million. At a thousand orders in a day the chance of
at least one collision is around three per cent; at five thousand it is a coin
flip; at ten thousand it is near certain. The suffix is now ten characters - 40
bits - which stays under a rounding error past a million orders a day. The cost
is four characters on a number people read out.

The failure was at least in the safe order: TryPlaceAsync runs before the
payment is charged, so a collision cost the customer an error rather than money.

A test pins the width so it cannot be shortened again for tidiness.

Cart quantity wrapped instead of clamping. AddCartItemHandler summed the
existing line and the requested amount in int arithmetic before clamping to
available stock. A quantity near int.MaxValue wrapped negative, Math.Min then
chose the negative, and the shopper was told the item was out of stock - a
misleading answer rather than a dangerous one, since nothing negative reached
the cart. The sum is now widened to long before the clamp, so the clamp does
the clamping and five available means five in the cart.

482 backend tests pass locally against PostgreSQL 16; dotnet format clean.

+++ src/WidgetWorks.Application/Carts/AddItem/AddCartItemHandler.cs
@@ -24,7 +24,11 @@ public sealed class AddCartItemHandler(ICartRepository carts, IWidgetRepository 
 
         var cart = await ResolveCartAsync(command.CartId, command.UserId, ct);
         var existing = cart.Items.FirstOrDefault(i => i.WidgetId == command.WidgetId);
-        var desired = Math.Min((existing?.Quantity ?? 0) + command.Quantity, widget.QuantityAvailable);
+        // Widened to long before adding. In int arithmetic a quantity near int.MaxValue wraps
+        // negative, Math.Min then picks the negative, and the shopper is told the item is out of
+        // stock - a misleading answer to a bad request rather than a dangerous one, but the clamp
+        // should do the clamping.
+        var desired = (int)Math.Min((long)(existing?.Quantity ?? 0) + command.Quantity, widget.QuantityAvailable);
         if (desired <= 0)
         {
             return Result<CartView>.Fail("This widget is out of stock.");
